requirements.txt
matplotlib==3.4.3
numpy==1.20.3
Pillow==8.4.0
pydicom==2.3.1
PyQt5==5.15.9
PyQt5_sip==12.9.0
